Few changes to make Stripe work. Probably not finished yet, but commited so I could pull changes from main.

+++ lib/data/repositories/coupon_repository.dart
@@ -128,9 +128,9 @@ class CouponRepository {
         try {
           coupons.add(Coupon(
             id: doc.id,
-            reduction: doc['reduction'],
+            reduction: doc['reduction'].toDouble(),
             reductionIsPercentage: doc['reductionIsPercentage'],
-            price: doc['pricePLN'],
+            price: doc['pricePLN'].toDouble(),
             hasLimits: doc['hasLimits'],
             worksOnline: doc['worksOnline'],
             worksInStore: doc['worksInStore'],
@@ -267,9 +267,9 @@ class CouponRepository {
 
     return Coupon(
       id: doc.id,
-      reduction: doc['reduction'],
+      reduction: doc['reduction'].toDouble(),
       reductionIsPercentage: doc['reductionIsPercentage'],
-      price: doc['pricePLN'],
+      price: doc['pricePLN'].toDouble(),
       hasLimits: doc['hasLimits'],
       worksOnline: doc['worksOnline'],
       worksInStore: doc['worksInStore'],
